requirements.txt
PyMySQL==0.10.0
Django==1.11.29
beautifulsoup4==4.9.1
Pillow==7.2.0
